update ohmyzsh install/plugins

+++ makesymlinks.sh
@@ -6,8 +6,9 @@
 
 ########## Variables
 
-dir=~/dotfiles                    # dotfiles directory
-olddir=~/dotfiles_old             # old dotfiles backup directory
+dir=~/dotfiles                              # dotfiles directory
+olddir=~/dotfiles_old                       # old dotfiles backup directory
+zshcustomsdir=~/.oh-my-zsh/custom/          # ohmyzsh plugins directory
 files="bashrc vimrc vim zshrc oh-my-zsh"    # list of files/folders to symlink in homedir
 
 ##########
@@ -24,10 +25,10 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if [[ -f ~/.$file ]];then
+        echo "Moving any existing dotfiles from ~ to $olddir"
+        mv ~/.$file ~/dotfiles_old/
+    fi
 done
 
 install_oh_my_zsh () {
@@ -35,7 +36,17 @@ install_oh_my_zsh () {
 if [ -f /bin/zsh -o -f /usr/bin/zsh ]; then
     # Clone my oh-my-zsh repository from GitHub only if it isn't already present
     if [[ ! -d $dir/oh-my-zsh/ ]]; then
+        export ZSH=$dir/oh-my-zsh/ sh install.sh
         sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+    fi
+    if [[ ! -d $zshcustomsdir/plugins/zsh-autosuggestions ]]; then
+        git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+    fi
+    if [[ ! -d $zshcustomsdir/plugins/zsh-syntax-highlighting ]]; then
+        git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+    fi
+    if [[ ! -d $zshcustomsdir/themes/powerlevel10k ]]; then
+        git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
     fi
     # Set the default shell to zsh if it isn't currently set to zsh
     if [[ ! $(echo $SHELL) == $(which zsh) ]]; then
@@ -45,3 +56,14 @@ fi
 }
 
 install_oh_my_zsh
+
+# move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
+for file in $files; do
+    if [[ -f ~/.$file ]];then
+        echo "Moving any existing dotfiles from ~ to $olddir"
+        mv ~/.$file /tmp/
+    fi
+
+    echo "Creating symlink to $file in home directory."
+    ln -s $dir/$file ~/.$file
+done
